% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_single_sample_util.R
\name{bisection_search}
\alias{bisection_search}
\title{helper function for direct_optimise}
\usage{
bisection_search()
}
\description{
helper function for direct_optimise
}
\details{
See direct_optimise.
}
\keyword{internal}
