% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_proposals.R
\name{cross_validate_replicates}
\alias{cross_validate_replicates}
\title{Cross validation using likelihood function}
\usage{
cross_validate_replicates(fct_list, split_method = "leave_one_out",
  lambda_vec = c(0, 2^(-5:5), 4096), starts = expand.grid(alpha =
  c(100, 0.01), delta = c(100, 0.01, 0.001, 1e-05)),
  use_pen_likelihood = F, verbose = F, print_details_to_file = T,
  k = 0, out_dir = paste0(getwd(), "/sim_details"), ...)
}
\arguments{
\item{fct_list}{A list of frequency count tables, assumed to be replicates.}

\item{split_method}{The method for choosing test and training data in each split.}

\item{lambda_vec}{The values of the penalty parameter we will train over.}

\item{starts}{Starting values for \code{alpha} and \code{delta} in the MLE procedure.}

\item{use_pen_likelihood}{Logical for whether we will use the penalized likelihood when evaluating on the test data in each split.}

\item{verbose}{logical.  If true then we additionally return a data frame with the details from each split.}

\item{print_details_to_file}{Prints the verbose output to a csv file.  For use during simulations if the split details are of potential interest.}

\item{k}{The number of splits if \code{split_method = "k_splits"}.}

\item{out_dir}{The file output for print_details_to_file}
}
\description{
\code{cross_validate_replicates} is a function which finds the maximum likelihood solution over a set of penalty parameter values, then determines which is best by evaluating the likelihood on the test data.
}
\details{
Coming soon.  Note to self:  use_pen_likelihood, out_dir and print_details_to_file should be removed before this is released to anyone else.
}
