% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_single_sample_util.R
\name{direct_optimise}
\alias{direct_optimise}
\title{Maximum likelihood estimate for single frequency count table}
\usage{
direct_optimise(fct, penalty = NULL, lambda = NULL,
  search_scheme = "grid", multiplier = 10, c_seq_len = 100,
  starts = NULL, alpha_min = 1e-10, delta_min = 1e-10,
  alpha_max = 1e+05, delta_max = 1e+05, forced_ccc_lower_bound = NA)
}
\arguments{
\item{fct}{a frequency count table, see \code{\link{make_frequency_count_table}} for more information.}

\item{penalty}{method of penalizing/regularizing the maximum likelihood solution.}

\item{lambda}{parameter for the h1 penalization method.}

\item{search_scheme}{method for iterating over candidate C values.}

\item{multiplier}{Richness values from the observed richness to \code{multiplier}*(observed richness) will be tested (default: 10).}

\item{c_seq_len}{Number of candidate C values we will maximize over.}

\item{starts}{A dataframe or matrix of optimization starts.  See details.}

\item{alpha_min, delta_min, alpha_max, delta_max}{box constraints for optimization over \code{alpha} and \code{delta}.}

\item{forced_ccc_lower_bound}{allows the user to force a lower bound to the grid search for C.}
}
\description{
Finds the MLE for a single frequency count table with the option of several search schemes and penalization choices.
}
\details{
THIS STILL NEEDS DETAILS, WORK IN PROGRESS.
}
\examples{
direct_optimise(nb_fct_simulation(1000, 100, 0.95))

# Suppose we want something quicker:
direct_optimise(nb_fct_simulation(1000, 100, 0.95),
    search_scheme = "bisection",
    c_seq_len = 20)

}
