% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_util.R
\name{direct_optimise_replicates}
\alias{direct_optimise_replicates}
\title{Maximum likelihood estimate for a list of frequency count tables.}
\usage{
direct_optimise_replicates(fct_list, penalty = NULL, lambda = NULL,
  search_scheme = "grid", multiplier = 10, c_seq_len = 100,
  starts = NULL, alpha_min = 1e-10, delta_min = 1e-10,
  alpha_max = 1e+05, delta_max = 1e+05)
}
\arguments{
\item{fct_list}{a list of frequency count tables, see \code{\link{make_frequency_count_table}} for more information on this data structure.}

\item{penalty}{method of penalizing/regularizing the maximum likelihood solution.}

\item{lambda}{parameter for the h1 penalization method.}

\item{search_scheme}{method for iterating over candidate C values.}

\item{multiplier}{Richness values from the observed richness to \code{multiplier}*(observed richness) will be tested (default: 10).}

\item{c_seq_len}{Number of candidate C values we will maximize over.}

\item{starts}{A dataframe or matrix of optimization starts.  See details.}

\item{alpha_min}{box constraints for optimization over \code{alpha} and \code{delta}.}

\item{delta_min}{box constraints for optimization over \code{alpha} and \code{delta}.}

\item{alpha_max}{box constraints for optimization over \code{alpha} and \code{delta}.}

\item{delta_max}{box constraints for optimization over \code{alpha} and \code{delta}.}
}
\description{
Finds the MLE for a list of replicate frequency count tables
  (i.e. frequency count tables from the same ecosystem) with the option of
  several search schemes and penalization choices.
}
\details{
THIS STILL NEEDS MUCHO DETAILS, WORK IN PROGRESS.
}
\examples{
list_of_fct <- list(nb_fct_simulation(1000, 0.1, 0.1),
                    nb_fct_simulation(1000, 0.1, 0.1)) # two replicates
direct_optimise_replicates(list_of_fct)

}
