% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_functions.R
\name{fct_to_observations}
\alias{fct_to_observations}
\title{Creates a vector of observations from a frequency count table}
\usage{
fct_to_observations(fct)
}
\arguments{
\item{fct}{A data frame in the form of a frequency count table.}
}
\description{
Takes a frequency count table and outputs the observations for
  each speices observed at least once in the data set.
}
\details{
Species observed zero times cannot be included in the output of this
  function because we have established a convention that a FCT does not
  contain that information.  This function is the inverse of
  \code{\link{make_frequency_count_table()}} if all species are observed
  at least once.
}
\examples{
toy_fct <- data.frame(freq = c(1,2,3), numTaxa = c(14,7,2))
abun <- fct_to_observations(toy_fct) # abundances for observed taxa
fct <- make_frequency_count_table(abun) # back to where we started
}
\seealso{
\code{\link{make_frequency_count_table()}}
}
