% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_proposals.R
\name{fixed_lambda_mle}
\alias{fixed_lambda_mle}
\title{Fixed lambda MLE for supplemental simulations}
\usage{
fixed_lambda_mle(fct_list, starts = data.frame(alpha = c(0.01, 0.01),
  delta = c(0.01, 1e-04)), lambda = 0, ...)
}
\arguments{
\item{fct_list}{A list of frequency count tables, assumed to be biological replicates.}

\item{starts}{Starting values for \code{alpha} and \code{delta} in the MLE procedure.}

\item{lambda}{The fixed lambda value to run the MLE procedure with.}
}
\description{

}
\details{

}
