% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_functions.R
\name{get_cc_ks_fs}
\alias{get_cc_ks_fs}
\title{Get formatted frequency count table data}
\usage{
get_cc_ks_fs(fct)
}
\arguments{
\item{fct}{A frequency count table.}
}
\value{
This function returns a list with cc (observed richness), and vectors which are just the columns of the frequency count table.
}
\description{
Formats the frequency count table data in the format
  required by the liklihood function
}
\details{
I may hide this from the user in the final version of this package.  It is primarily useful for running tests on the likelihood function.
}
\examples{
toy_fct <- data.frame(freq = c(1,2,3), numTaxa = c(14,7,2))
get_cc_ks_fs(toy_fct)
}
