% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_functions.R
\name{get_cc_max}
\alias{get_cc_max}
\title{Returns the maximum observed richness for a FCT list.}
\usage{
get_cc_max(list_of_fct)
}
\description{
Returns the maximum observed richness for a FCT list.
}
\keyword{internal}
