% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_single_sample_util.R
\name{log_marginal_negative_binomial}
\alias{log_marginal_negative_binomial}
\title{Helper for likelihood.}
\usage{
log_marginal_negative_binomial(k, alpha, delta)
}
\description{
Helper for likelihood.
}
\details{
Helper for likelihood.
}
\keyword{internal}
