% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_single_sample_util.R
\name{loglike_single_fixed_ccc}
\alias{loglike_single_fixed_ccc}
\title{Likelihood function written in R for reference}
\usage{
loglike_single_fixed_ccc(x, ccc, cc, ks, fs, penalty, lambda)
}
\description{
Likelihood function written in R for reference
}
\details{
Log likelihood function, completely equivalent to the C++ version which is actually used for computation.
}
\keyword{internal}
