% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_functions.R
\name{make_frequency_count_table}
\alias{make_frequency_count_table}
\title{Creates a frequency count table}
\usage{
make_frequency_count_table(abundances, cautious = T)
}
\arguments{
\item{abundances}{A numeric vector of absolute species abundances.}

\item{cautious}{logical.  If TRUE (default) then a vector of all zeroes will
return a data frame.  If FALSE then an error will be thrown.}
}
\value{
returns a data.frame.
}
\description{
\code{make_frequency_count_table()} takes a vector of species abundances and
returns a frequency count table.
}
\details{
A frequency count table is a data frame listing the frequency ('freq') and
the number of species observed that many times in the sample ('numTaxa').

In order to allow this function to be used easily with simulated data, we
omit any zero observations from our frequency count table.

If \code{cautious = T}, in the event that all of the species abundances
listed are zero, the function returns a data frame with the correct labels
for a frequency count table but zero rows.
}
\examples{
taxa <- rnbinom(1000,0.1,0.1) # simulated data
make_frequency_count_table(taxa)

\dontrun{
# The following will produce an error:
make_frequency_count_table(rep(0,100),cautious= F)
}
}
