% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_proposals.R
\name{min_var_C_hat}
\alias{min_var_C_hat}
\title{Tuning penalty parameter by minimum variance criterion}
\usage{
min_var_C_hat(replicate_list, lambda_vec, rtn_full_list = F, ...)
}
\arguments{
\item{replicate_list}{A list of frequency count tables}

\item{rtn_full_list}{logical.  If F we only return the best lambda and the
corresponding estimate.  If T we return all lambdas and estimates in a
dataframe.}

\item{lamdba_vec}{A set of lambda values which we propose testing over.}
}
\description{
This function chooses the \code{lambda} value which produces C
  estimates with the smallest variance.
}
\details{
We found by simulation that this consistently chooses the largest
  \code{lambda} we test.  Therefore this function is not recommended, but we
  keep it for organizing our ideas.
}
