% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_functions.R
\name{nb_fct_simulation}
\alias{nb_fct_simulation}
\title{Negative binomial frequency count table simulation}
\usage{
nb_fct_simulation(C, alpha, delta)
}
\arguments{
\item{C}{The true species richness}

\item{alpha}{Shape parameter in the gamma distribution (our assumed
mixing distribution)}

\item{delta}{Rate parameter in the gamma distribution (our assumed
mixing distribution)}
}
\value{
Returns a data frame in the form of a frequency count table
}
\description{
Simulates a frequency count table according to the negative
  binomial model with the parameters input.
}
\details{
Assuming a gamma mixed Poisson model, the number of times any
  taxa is observed follows a single draw from the negative binomial
  distribution.  To simulate for C species, we make C draws.
}
\examples{
nb_fct_simulation(1000, 0.1, 0.001) #output is a frequency count table
}
