% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_proposals.R
\name{one_split}
\alias{one_split}
\title{Helper function for \code{cross_validate_replicates}}
\usage{
one_split(pen_likelihood = T)
}
\description{
Helper function for \code{cross_validate_replicates}
}
\keyword{internal}
