% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty_single_sample_util.R
\name{optimise_fixed_c}
\alias{optimise_fixed_c}
\title{helper function called by grid_search and bisection_search}
\usage{
optimise_fixed_c()
}
\description{
helper function, see direct_optimise.
}
\details{
See direct_optimise.
}
\keyword{internal}
