% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_util.R
\name{replicate_likelihood}
\alias{replicate_likelihood}
\title{gamma-Poisson log likelihood for replicates}
\usage{
replicate_likelihood(x, ccc, cc_list, ks_list, fs_list, penalty, lambda)
}
\arguments{
\item{x}{a length 2 vector, \eqn{eta = (\alpha, \delta)}.}

\item{ccc}{The value of \eqn{C} to evaluate the likelihood at.}

\item{cc_list}{A list, each item is a length 1 numeric vector stating the observed richness for each replicate.}

\item{ks_list}{A list, each list item is a vector for that replicate.  The
vector represents the \eqn{k} values for the nonzero frequencies.}

\item{fs_list}{A list, each list item is a vector for that replicate.  The
vector represents the frequencies \eqn{f_k} associated with the \eqn{k}
values in \code{ks_list}.}

\item{penalty}{The penalty function form to use, only "h1" is supported currently.}

\item{lambda}{The penalization parameter.}
}
\description{
The gamma-Poisson likelihood (which has parameters \eqn{C, \eta}) for a sample of frequency count tables drawn from the same population.
}
\details{
The paramaters \code{cc_list}, \code{ks_list}, \code{fs_list} are a
  formatted version of the frequency count table list.  The likelihood
  function is called many times as we optimise, so we require them to be
  formatted this way to save time in the long run.  This function uses RCPP
  to speed up the unavoidable factorial calculations in the likelihood.
}
\examples{
list_of_fct <- rre::nb_fct_simulation(1000, 0.1, 0.1, 2) # two replicates
direct_optimise_replicates(list_of_fct)

}
\keyword{internal}
