% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_proposals.R
\name{unregularized_mle}
\alias{unregularized_mle}
\title{Proposal 0: Unregularized MLE wrapper for simulations}
\usage{
unregularized_mle(fct_list, starts = data.frame(alpha = c(0.01, 0.01),
  delta = c(0.01, 1e-04)), ...)
}
\arguments{
\item{fct_list}{A list of frequency count tables, assumed to be biological
replicates.}

\item{starts}{Starting values for \code{alpha} and \code{delta} in the MLE
procedure.}
}
\description{
Maximum likelihood estimate without regularization.
}
\details{
This is used as the comparison point for our tuning proposals, a
  wrapper to conform to simulation output.
}
