% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_functions.R
\name{ztnb_ev}
\alias{ztnb_ev}
\alias{ztnb_var}
\title{Expected value and variance of the zero truncated negative binomial}
\usage{
ztnb_ev(alpha, p = NULL, delta = NULL)

ztnb_var(alpha, p = NULL, delta = NULL)
}
\arguments{
\item{alpha}{The \code{size} parameter fo \code{\link[stats]{rnbinom}} (also
the shape parameter in the gamma mixing distribution, see 'Details')}

\item{p}{The \code{prob} parameter of \code{\link[stats]{rnbinom}}.}

\item{delta}{An alternative parametrization to using p.  (see 'Details')}
}
\value{
A length 1 numeric vector.
}
\description{
Functions for calculating the mean or variance of the zero truncated negative binomial
  distribution (two parametrizations available).
}
\details{
If \code{alpha} and \code{p} are passed, then we follow the
  parameterization used in \code{\link[stats]{rnbinom}}.  If \code{alpha} and
  \code{delta} are passed, then the distribution follows the parametrization
  derived by assuming the number of observations for each species is a
  gamma-mixed Poisson.  In this derivation the abundances are
  \eqn{Gamma(\alpha, \delta)} in the shape/rate parametrization.

  These parametrizations are related by \eqn{p = \delta  / (1+\delta)}.

  If \code{p} and \code{delta} are both passed then an error is thrown.
}
\examples{
ztnb_ev(0.1,0.5) # using p
identical(ztnb_ev(0.1, delta = 1), ztnb_ev(0.1,0.5)) # equivalent.

ztnb_var(100,0.95)
}
