% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_proposals.R
\name{cv_replicates}
\alias{cv_replicates}
\title{Proposal 4: Cross validation for data with replicates}
\usage{
cv_replicates(fct_list, lambda_vec = seq(0, 20, by = 2),
  starts = data.frame(alpha = c(0.01, 0.01), delta = c(0.01, 1e-04)),
  partitions = 10, eval_function = "gof_chi_sq", ...)
}
\arguments{
\item{fct_list}{A list of frequency count tables, assumed to be replicates.}

\item{lambda_vec}{The values of the penalty parameter we will train over.}

\item{starts}{Starting values for \code{alpha} and \code{delta} in the MLE procedure.}

\item{partitions}{An integer indicating the number of times to randomly split the data into testing and validating subsets.}

\item{eval_function}{A function which evaluates how well a set of parameters fit a list of frequency count tables.  To conform to goodness of fit, we will assume that lower scores are better.}
}
\description{
Randomly partitions the data into training and evaluation subsets and selects the best parameter estimates based on various evaluation metrics.
}
\details{
For each partition the we find the MLE solution, repeating for a grid of regularization values given by \code{lambda_vec}.  We then evaluate how well these parameters fit using the training data and a choice of evaluation functions.  Currently we support goodness of fit (chi-square) and unregularized likelihood as possible evaluation functions.
}
