% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_proposals.R
\name{gof_criterion}
\alias{gof_criterion}
\title{Goodness of fit regularization criterion}
\usage{
gof_criterion(fct_list, lambda_vec = seq(0, 20, by = 2),
  starts = data.frame(alpha = c(0.01, 0.01), delta = c(0.01, 1e-04)),
  gof_method = "chi_sq", ...)
}
\arguments{
\item{fct_list}{A list of frequency count tables, assumed to be biological replicates.}

\item{lambda_vec}{The values of the penalty parameter we will train over.}

\item{starts}{Starting values for \code{alpha} and \code{delta} in the MLE procedure.}

\item{gof_method}{Goodness of fit function, currently chi^2 is the only option.}
}
\description{
Optimizes with a grid of regularization parameters and chooses the best one via goodness of fit metrics.
}
\details{
Currently only chi^2 is supported, but there are potentially better goodness of fit metrics for count data.
}
