% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_proposals.R
\name{minimum_subset_distance}
\alias{minimum_subset_distance}
\title{Minimum subset distance}
\usage{
minimum_subset_distance(fct_list, lambda_vec = seq(0, 20, by = 2),
  starts = data.frame(alpha = c(0.01, 0.01), delta = c(0.01, 1e-04)),
  partitions = 10, ...)
}
\arguments{
\item{fct_list}{A list of frequency count tables, assumed to be replicates.}

\item{lambda_vec}{The values of the penalty parameter we will train over.}

\item{starts}{Starting values for \code{alpha} and \code{delta} in the MLE procedure.}

\item{partitions}{An integer indicating the number of times to randomly split the data into testing and validating subsets.}
}
\description{
Paritions the frequency count table list into two subsets and
  finds the lambda value which minimizes the variance (distance) between them.
}
\details{
For each round of we partition the data into two subsets.  The MLE for species richness is found in each, using all lambda in lambda_vec.  The lambda which allowed the lowest variance between subset MLEs is selected as best, and we report the average of ccc, alpha and delta between those two.
}
